06032025 - Update Delivery Acceptance Controller and FXML
- Update value of quantity
- Update size of cost, order, receive and total in table fxml

+++ src/main/java/com/rmj/guanzongroup/sidebarmenus/controller/DeliveryAcceptance_EntryController.java
@@ -2070,7 +2070,7 @@ public class DeliveryAcceptance_EntryController implements Initializable, Screen
                                             String.valueOf(poPurchaseReceivingController.Detail(lnCtr).Inventory().getDescription()),
                                             String.valueOf(CustomCommonUtil.setIntegerValueToDecimalFormat(poPurchaseReceivingController.Detail(lnCtr).getUnitPrce(), true)),
                                             String.valueOf(poPurchaseReceivingController.Detail(lnCtr).getOrderQty().intValue()),
-                                            String.valueOf(poPurchaseReceivingController.Detail(lnCtr).getQuantity()),
+                                            String.valueOf(poPurchaseReceivingController.Detail(lnCtr).getQuantity().intValue()),
                                             String.valueOf(CustomCommonUtil.setIntegerValueToDecimalFormat(lnTotal, true)) //identify total
                                     ));
                         }
